Move volume button utils to seperate file
